refactor(AfterSignInScreen): Merge with AfterSigninSwitch. Updated various comments and reorg import
