+++ lib/actions/ui.js
@@ -11,11 +11,13 @@ import { setActiveItinerary } from './narrative'
 import { getUiUrlParams } from '../util/state'
 
 /**
- * Wrapper function for history#push that preserves the current search or, if
+ * Wrapper function for history#push (or, if specified, replace, etc.)
+ * that preserves the current search or, if
  * replaceSearch is provided (including an empty string), replaces the search
  * when routing to a new URL path.
  * @param  {[type]} url           path to route to
  * @param  {string} replaceSearch optional search string to replace current one
+ * @param  {func}   routingMethod the connected-react-router method to execute (defaults to push).
  */
 export function routeTo (url, replaceSearch, routingMethod = push) {
   return function (dispatch, getState) {
