+++ lib/actions/user.js
@@ -1,6 +1,7 @@
 import { createAction } from 'redux-actions'
 
 import { addUser, fetchUser, updateUser } from '../util/middleware'
+import { isNewUser } from '../util/user'
 
 const setCurrentUser = createAction('SET_CURRENT_USER')
 
@@ -16,10 +17,6 @@ function getStateForNewUser (auth0User) {
     savedLocations: [],
     storeTripHistory: false // User must opt in.
   }
-}
-
-export function isNewUser (loggedInUser) {
-  return !loggedInUser.hasConsentedToTerms
 }
 
 /**
