+++ lib/components/app/responsive-webapp.js
@@ -7,17 +7,16 @@ import React, { Component } from 'react'
 import { connect } from 'react-redux'
 import { Route, Switch, withRouter } from 'react-router'
 
+import PrintLayout from './print-layout'
 import { setMapCenter, setMapZoom } from '../../actions/config'
 import { formChanged, parseUrlQueryString } from '../../actions/form'
 import { getCurrentPosition, receivedPositionResponse } from '../../actions/location'
 import { setLocationToCurrent } from '../../actions/map'
 import { handleBackButtonPress, matchContentToUrl } from '../../actions/ui'
 import { getActiveItinerary, getTitle } from '../../util/state'
-import AfterSignInSwitch from '../user/after-signin-switch'
+import AfterSignInScreen from '../user/after-signin-screen'
 import UserAccountScreen from '../user/user-account-screen'
 import withLoggedInUserSupport from '../user/with-logged-in-user-support'
-
-import PrintLayout from './print-layout'
 
 const { isMobile } = coreUtils.ui
 
@@ -217,13 +216,14 @@ class RouterWrapper extends Component {
               }}
             />
             <Route
-              // This route is called immediately after login
-              // and also by the auth0 redirect callback.
-              // The component for this route determines what is ultimately shown after login.
+              // This route is called immediately after login by Auth0
+              // and by the onRedirectCallback function from /lib/util/auth.js.
+              // For new users, it displays the account setup form.
+              // For existing users, it takes the browser back to the itinerary search prior to login.
               path={'/signedin'}
               component={routerProps => {
                 const props = this._combineProps(routerProps)
-                return <AfterSignInSwitch {...props} />
+                return <AfterSignInScreen {...props} />
               }}
             />
             <Route
