+++ lib/components/user/user-account-screen.js
@@ -4,9 +4,9 @@ import { connect } from 'react-redux'
 import { withLoginRequired } from 'use-auth0-hooks'
 
 import { routeTo } from '../../actions/ui'
-import { addOrUpdateUser, isNewUser } from '../../actions/user'
+import { addOrUpdateUser } from '../../actions/user'
+import { isNewUser } from '../../util/user'
 import AppNav from '../app/app-nav'
-
 import AccountSetupFinishPane from './account-setup-finish-pane'
 import ExistingAccountDisplay from './existing-account-display'
 import FavoriteLocationsPane from './favorite-locations-pane'
