+++ lib/util/auth.js
@@ -66,16 +66,11 @@ export function getAuth0Callbacks (reduxStore) {
      */
     onRedirectCallback: appState => {
       if (appState && appState.urlHash) {
-        // Immediately after successful login, Auth0 will redirect to
-        // /signedin (Auth0-whitelisted) with no search parameters.
-        // This code block adds the the hash parameters to /signedin,
-        // so that if the user has set up their MOD UI account, the handler for /signedin
-        // can route the browser to the itinerary search just before login (stored in appState.urlHash).
-        // This code block prevents the URL with no search from appearing in the user's browsing history.
-        const newUserPath = appState.urlHash.replace('#/', '#/signedin')
-        const newUserUrl = `${URL_ROOT}${newUserPath}`
-
-        window.location.replace(newUserUrl)
+        // Immediately after successful login, Auth0 redirects to /signedin (Auth0-whitelisted).
+        // Here, we add back the search query params that are contained within urlHash to preserve any itinerary search
+        // that may have been present before login (and avoid a URL without a search showing in the user's browsing history).
+        const nextUrl = `${URL_ROOT}${appState.urlHash.replace('#/', '#/signedin')}`
+        window.location.replace(nextUrl)
       } else if (appState && appState.returnTo) {
         // TODO: Handle other after-login situations.
         // Careful!
