+++ lib/util/user.js
@@ -1,0 +1,4 @@
+
+export function isNewUser (loggedInUser) {
+  return !loggedInUser.hasConsentedToTerms
+}
